requirements.txt
streamlit
torch
torchvision
Pillow
scikit-learn
